// src/types.ts
export interface PokemonList {
  count: number;
  next: string;
  previous: null;
  data: Result[];
}

export interface Result {
  id: number;
  identifier: string;
  species_id: number;
  height: number;
  weight: number;
  base_experience: BaseExperience;
  order_number: BaseExperience;
  is_default: number;
  url: string;
  details?: IPokemonDetail;
}

export interface BaseExperience {
  Int64: number;
  Valid: boolean;
}

export interface Result {
  id: number;
  identifier: string;
  species_id: number;
  height: number;
  weight: number;
  base_experience: BaseExperience;
  order_number: BaseExperience;
  is_default: number;
  url: string;
}

export interface BaseExperience {
  Int64: number;
  Valid: boolean;
}

export interface IPokemonDetail {
  abilities: Ability[];
  base_experience: number;
  cries: Cries;
  forms: Species[];
  game_indices: GameIndex[];
  height: number;
  held_items: any[];
  id: number;
  is_default: boolean;
  location_area_encounters: string;
  moves: Move[];
  name: string;
  order: number;
  past_abilities: any[];
  past_types: any[];
  species: Species;
  sprites: Sprites;
  stats: Stat[];
  types: Type[];
  weight: number;
  url: string;
}

export interface Ability {
  ability: Species;
  is_hidden: boolean;
  slot: number;
}

export interface Species {
  name: string;
  url: string;
}

export interface Cries {
  latest: string;
  legacy: string;
}

export interface GameIndex {
  game_index: number;
  version: Species;
}

export interface Move {
  move: Species;
  version_group_details: VersionGroupDetail[];
}
// ...
export interface Sprites {
  back_default: string;
  back_female: null;
  back_shiny: string;
  back_shiny_female: null;
  front_default: string;
  front_female: null;
  front_shiny: string;
  front_shiny_female: null;
  other?: Other;
  versions?: Versions;
  animated?: Sprites;
}
// ...
export interface Stat {
  base_stat: number;
  effort: number;
  stat: Species;
}

export interface Type {
  slot: number;
  type: Species;
}